+++ ts/features/itwallet/presentation/proximity/utils/errors.ts
