+++ ts/features/itwallet/presentation/proximity/utils/types.ts
